Database: methods for saving metrics split to SaveMetricsClient
